zephyr: Move lvgl input driver to separate file
Move the LVGL pointer input driver initialization code to
lvgl_input_sink file. Also prepare the code for direct usage of the
input subsystem instead of proxying it through the kscan handler.
