+++ zephyr/Kconfig.input
@@ -4,38 +4,58 @@
 
 menu "Input device settings"
 
-config LV_Z_POINTER_KSCAN
-	bool "Keyboard scan pointer input"
-	depends on KSCAN
+
+choice LV_Z_POINTER_DRIVER
+	prompt "Pointer input driver"
+	default LV_Z_POINTER_NONE
 	help
-	  Enable keyboard scan pointer input
+	  Specify the lvgl pointer driver implementation.
 
-if LV_Z_POINTER_KSCAN
+	config LV_Z_POINTER_NONE
+	  bool "Disable lvgl pointer"
 
-config LV_Z_POINTER_KSCAN_MSGQ_COUNT
-	int "Keyboard scan message queue count maximum"
+	config LV_Z_POINTER_INPUT
+	  bool "Input Subsystem"
+	  depends on INPUT
+	  select LV_Z_POINTER
+
+	config LV_Z_POINTER_KSCAN
+	  bool "Keyboard scan"
+	  depends on KSCAN
+	  select LV_Z_POINTER
+
+
+endchoice
+
+config LV_Z_POINTER
+	bool
+
+if LV_Z_POINTER
+
+config LV_Z_POINTER_DRIVER_MSGQ_COUNT
+	int "Pointer message queue count maximum"
 	default 10
 	help
-	  Maximum number of items in the keyboard scan message queue.
+	  Maximum number of items in the pointer message queue.
 
-config LV_Z_POINTER_KSCAN_SWAP_XY
+config LV_Z_POINTER_DRIVER_SWAP_XY
 	bool "Swap keyboard scan X,Y axes"
 	help
-	  Swap keyboard scan X,Y axes. This option can be used to align keyboard
-	  scan coordinates with the display.
+	  Swap pointer X,Y axes. This option can be used to align
+	  coordinates with the display.
 
-config LV_Z_POINTER_KSCAN_INVERT_X
+config LV_Z_POINTER_DRIVER_INVERT_X
 	bool "Invert keyboard scan X axis"
 	help
-	  Invert keyboard scan X axis. This option can be used to align keyboard
-	  scan coordinates with the display.
+	  Invert pointer X axis. This option can be used to align
+	  coordinates with the display.
 
-config LV_Z_POINTER_KSCAN_INVERT_Y
+config LV_Z_POINTER_DRIVER_INVERT_Y
 	bool "Invert keyboard scan Y axis"
 	help
-	  Invert keyboard scan Y axis. This option can be used to align keyboard
-	  scan coordinates with the display.
+	  Invert pointer scan Y axis. This option can be used to align
+	  coordinates with the display.
 
-endif # LV_Z_POINTER_KSCAN
+endif # LV_Z_POINTER
 
 endmenu
